[refactor] - 디렉토리명 변경 - 구조 변경 - Controller를 Service 결과 값을 View에 연결 하는 브로커 역할로 수정
