+++ src/test/java/study/step3/CarTest.java
@@ -4,8 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
-import step3.model.Car;
+import step3.Model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
